# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Async payment processing service",
        "title": "Payment Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/payments": {
            "post": {
                "description": "Initiates a payment transaction and sends it to the processing system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Send payment to processing",
                "parameters": [
                    {
                        "description": "Payment request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/inport.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/inport.TXConfirmation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/{txid}": {
            "get": {
                "description": "Returns current status and details of a payment by TXID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get payment status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "txid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/inport.PaymentInfo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "inport.PaymentAccountDTO": {
            "type": "object",
            "properties": {
                "accCurency": {
                    "type": "string"
                },
                "iban": {
                    "type": "string"
                }
            }
        },
        "inport.PaymentInfo": {
            "type": "object",
            "properties": {
                "creditorAccount": {
                    "$ref": "#/definitions/inport.PaymentAccountDTO"
                },
                "debitorAccount": {
                    "$ref": "#/definitions/inport.PaymentAccountDTO"
                },
                "metadata": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "txid": {
                    "type": "string"
                }
            }
        },
        "inport.PaymentPartyDTO": {
            "type": "object",
            "properties": {
                "bic": {
                    "type": "string"
                },
                "contryOfResidence": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "inport.PaymentRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "creditor": {
                    "$ref": "#/definitions/inport.PaymentPartyDTO"
                },
                "creditorPacc": {
                    "$ref": "#/definitions/inport.PaymentAccountDTO"
                },
                "currency": {
                    "type": "string"
                },
                "debitor": {
                    "$ref": "#/definitions/inport.PaymentPartyDTO"
                },
                "debitorPacc": {
                    "$ref": "#/definitions/inport.PaymentAccountDTO"
                },
                "endToEndIdentification": {
                    "type": "string"
                },
                "metadata": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "txid": {
                    "type": "string"
                }
            }
        },
        "inport.TXConfirmation": {
            "type": "object",
            "properties": {
                "metadata": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "txid": {
                    "type": "string"
                }
            }
        }
    }
}